Fix according to requested changes
