+++ src/main/java/at/hannibal2/skyhanni/config/Storage.java
@@ -1,6 +1,7 @@
 package at.hannibal2.skyhanni.config;
 
 import at.hannibal2.skyhanni.data.model.ComposterUpgrade;
+import at.hannibal2.skyhanni.features.dungeon.DungeonAPI;
 import at.hannibal2.skyhanni.features.fishing.trophy.TrophyRarity;
 import at.hannibal2.skyhanni.features.garden.CropAccessory;
 import at.hannibal2.skyhanni.features.garden.CropType;
@@ -436,7 +437,7 @@ public class Storage {
         public static class DungeonStorage {
 
             @Expose
-            public Map<String, Integer> bosses = new HashMap<>();
+            public Map<DungeonAPI.DungeonFloor, Integer> bosses = new HashMap<>();
         }
     }
 }
